pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Backend Template"
authors = ["Danila Safonov <danil.sapog19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.5.2"}
python-jose = "^3.3.0"
uvicorn = "^0.24.0.post1"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1.2"
asyncpg = "^0.29.0"
orjson = "^3.9.10"


[tool.poetry.group.tools.dependencies]
pre-commit = "^3.5.0"


[tool.poetry.group.lints.dependencies]
mypy = "^1.7.1"
ruff = "^0.1.7"
black = "^23.11.0"
isort = "^5.12.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
testcontainers = {extras = ["postgres"], version = "^3.7.1"}
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
combine_as_imports = true
remove_redundant_aliases = true
force_grid_wrap = 2
use_parentheses = true
ensure_newline_before_comments = true
